fix #270. 完成了一些简单的代码模块化工作, 进一步模块化工作均用TODO在文件开头标出.
