+++ src/controllers/projectInfo.go
@@ -2,6 +2,7 @@ package controllers
 
 import (
 	"Kcoin-Golang/src/models"
+	"Kcoin-Golang/src/service"
 
 	"github.com/astaxie/beego"
 )
@@ -27,13 +28,13 @@ func (c *ProjectInfoController) Get() {
 	//这里的session都没有对不同项目进行区分，后续应当还需要更改
 	starNum := c.GetSession("starNum")
 	if starNum == nil {
-		starNum = models.GetStarNum(fakeURL)
+		starNum = service.GetStarNum(fakeURL)
 		c.SetSession("starNum", starNum)
 	}
 
 	contributorsNum := c.GetSession("contributorsNum")
 	if contributorsNum == nil {
-		contributorsNum = models.GetContributorNum(fakeURL)
+		contributorsNum = service.GetContributorNum(fakeURL)
 		c.SetSession("contributorsNum", contributorsNum)
 	}
 	projectCC, _ := models.GetProjectsCC(id.(string))
