+++ src/models/getAllJoinedProjects.go
@@ -1,3 +1,4 @@
+// TODO 该文件并入k_user.go中
 package models
 
 import (
