+++ src/models/getCommiters.go
@@ -1,6 +1,6 @@
-package models 
-//api:https://api.github.com/repos/OS-ABC/Kcoin-Golang/commits
+package models
 
+//api:https://api.github.com/repos/OS-ABC/Kcoin-Golang/commits
 
 import (
 	"fmt"
@@ -8,16 +8,18 @@ import (
 	"net/http"
 )
 
+// TODO 该函数尚未完全完成, 需要完成获取某个项目的所有commiter信息
 //验证是否是项目commiter的接口，参数为项目名称和用户名称
-func checkAutho(userName string,programName string) string {
-//请求：获取此项目的commiter信息
-	var url_1 string = "https://api.github.com/repos/"+userName+"/"+programName+"/"+"commits"
-	client :=&http.Client{}
-	response,_:=client.Get(url_1)
+func checkAutho(userName string, programName string) string {
+	//请求：获取此项目的commiter信息
+	var url_1 string = "https://api.github.com/repos/" + userName + "/" + programName + "/" + "commits"
+	client := &http.Client{}
+	response, _ := client.Get(url_1)
 	defer response.Body.Close()
-	body,err_1:=ioutil.ReadAll(response.Body)
-	if err_1 != nil{
-		panic(err_1)}
+	body, err_1 := ioutil.ReadAll(response.Body)
+	if err_1 != nil {
+		panic(err_1)
+	}
 
 	fmt.Println(string(body))
 	return string(body)
@@ -29,4 +31,3 @@ func checkAutho(userName string,programName string) string {
 //	fmt.Println(info)
 //	return info
 //}
-
