+++ src/models/sendEmailToPotentialUsers.go
@@ -1,3 +1,4 @@
+// TODO 该文件不应该放在model层, 应该放在service层
 package models
 
 import (
